Makes changes as requested by CR
