+++ src/App.js
@@ -1,9 +1,20 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import './App.css';
+import { useDispatch } from 'react-redux';
+import { useEffect } from 'react';
 import Layout from './components/Layout';
 import { Rockets, Missions, MyProfile } from './routes';
+import { fetchRockets } from './redux/rockets/rocketsSlice';
+import { getMission } from './redux/missions/missionsSlice';
 
 function App() {
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(getMission());
+    dispatch(fetchRockets());
+  }, [dispatch]);
+
   return (
     <BrowserRouter>
       <Routes>
