+++ src/redux/store.js
@@ -1,4 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit';
+import { logger } from 'redux-logger';
 import rocketReducer from './rockets/rocketsSlice';
 import missionSlice from './missions/missionsSlice';
 
@@ -7,6 +8,7 @@ const store = configureStore({
     rockets: rocketReducer,
     missions: missionSlice,
   },
+  middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
 });
 
 export default store;
